compile

+++ src/enclave/App/App.cpp
@@ -1195,7 +1195,7 @@ Java_edu_berkeley_cs_rise_opaque_execution_SGXEnclave_EnclaveColumnSort(
     input_copy[i] = *(ptr + i);
   }
 
-  uint32_t *output_buffer = (uint32_t *) malloc(sizeof(uint32_t) * r);
+  uint8_t *output_buffer = (uint8_t *) malloc(sizeof(uint8_t) * r);
   uint32_t output_buffer_size;
 
   if (round == 0) {
